requirements.txt
Django==1.10.7
django-filter==1.0.2
django-simple-menu==1.2.1
django-tables2==1.5.0
